+++ lib/Delete/index.js
@@ -1,3 +1,4 @@
+// Delete this folder if you're not using it
 /* eslint-disable no-console */
 const { Client, GatewayIntentBits } = require('discord.js');
 const { init } = require('../../commands');
