+++ lib/app.js
@@ -1,3 +1,5 @@
+// delete this if you're not making an express app
+
 const express = require('express');
 const app = express();
