+++ lib/commands/init.js
