+++ lib/middleware/README.md
@@ -1,3 +1,1 @@
-# Middleware
-
-This folder will hold all your app's custom middleware. For instance, if you need to create some middleware to authenticate users, you can put the middleware function file (e.g. `ensureAuth.js`) in this directory.
+# Delete this folder if you're not using it.
