+++ lib/models/Creator.js
@@ -14,6 +14,7 @@ module.exports = class Creator {
     this.image_id = row.image_id;
   }
 
+  // i would call this getAll
   static async getCreators() {
     const { rows } = await pool.query(
       `
