+++ lib/models/User.js
@@ -10,6 +10,7 @@ module.exports = class User {
     this.user_id = row.user_id;
     this.content = row.content;
   }
+
   static async insert({ user_id, content }) {
     const { rows } = await pool.query(
       `
@@ -21,6 +22,9 @@ module.exports = class User {
     );
     return new User(rows[0]);
   }
+
+  // if you change this model to Joke, this method could be 
+  // called random or getRandoms
   static async getRandomJoke({ user_id }) {
     const { rows } = await pool.query(
       `
@@ -34,6 +38,8 @@ module.exports = class User {
     );
     return new User(rows[0]);
   }
+
+  // If you change this model to Joke, this method could be simply delete
   static async deleteUserJoke(id) { 
     const { rows } = await pool.query(
       `
@@ -45,6 +51,8 @@ module.exports = class User {
     );
     return new User(rows[0]);
   }
+  // if you change this model to Joke the this method could be 
+  // called getJokesByUserId
   static async getAllUserJokes({ user_id }) {
     const { rows } = await pool.query(
       `
