+++ lib/services/client.js
@@ -1,0 +1,9 @@
+const client = new Client({
+  intents: [
+    GatewayIntentBits.Guilds,
+    GatewayIntentBits.GuildMessages,
+    GatewayIntentBits.MessageContent
+  ]
+});
+
+module.exports = client;
